# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: A Dynamic DNS REST Service.
  description: The routes of the dynamic DNS API
  contact:
    name: Thilo Schwarz
    url: https://github.com/th-schwarz/DynDRest
    email: osp (at) thischwa.codes
  license:
    name: MIT Licence
    url: https://github.com/th-schwarz/DynDRest/blob/develop/LICENSE
  version: "0.4"
...
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/ips/{host}:
    get:
      tags:
      - api-controller
      summary: Determines the IP settings of the 'host' and returns it in a JSON object.
      operationId: fetchHostIpSetting
      parameters:
      - name: host
        in: path
        description: "The host, for which the IPs must be determined."
        required: true
        schema:
          type: string
          description: "The host, for which the IPs must be determined."
          example: mydyndns.domain.com
        example: mydyndns.domain.com
      - name: apiToken
        in: query
        description: "The 'apiToken', which must belong to the host'."
        required: true
        schema:
          type: string
          description: "The 'apiToken', which must belong to the host'."
      responses:
        "200":
          description: A JSON object with the IP settings of the 'host'.
          content:
            application/json:
              example:
                ipv4: 127.1.2.4
                ipv6: 2a03:4000:41:32:0:0:0:2
        "403":
          description: "If the 'apiToken' doesn't belong to the host, IP addresses\
            \ aren't valid or the remote IP couldn't determine."
        "404":
          description: If the 'host' isn't configured.
        "500":
          description: If the zone info fails.
    put:
      tags:
      - api-controller
      summary: "Updates the desired IP addresses of the 'host'. If both parameters\
        \ for IP addresses aren't set, an attempt is made to fetch the remote IP."
      operationId: updateHost
      parameters:
      - name: host
        in: path
        description: "The host, for which the IPs must be updated. It has to be a\
          \ full domain name."
        required: true
        schema:
          type: string
          description: "The host, for which the IPs must be updated. It has to be\
            \ a full domain name."
          example: mydyndns.domain.com
        example: mydyndns.domain.com
      - name: apiToken
        in: query
        description: The 'apiToken' to authenticate the changes of the IPs for this
          host.
        required: true
        schema:
          type: string
          description: The 'apiToken' to authenticate the changes of the IPs for this
            host.
      - name: ipv4
        in: query
        description: An IPv4 address.
        required: false
        schema:
          type: string
          description: An IPv4 address.
          examples:
          - 127.1.2.4
      - name: ipv6
        in: query
        description: An IPv6 address.
        required: false
        schema:
          type: string
          description: An IPv6 address.
          examples:
          - 2a03:4000:41:32:0:0:0:2
      responses:
        "200":
          description: The IPs are still up to date.
        "201":
          description: 'One or both IPs are changed. Update successful processed.
            (Dependent on the configuration the response code could be ''200''! '
        "400":
          description: At least one IP address isn't valid or the remote IP couldn't
            be determined.
        "403":
          description: If the 'apiToken' doesn't belong to the host.
        "404":
          description: The desired host doesn't exists.
        "500":
          description: If the addOrUpdate failed.
...
  /router/{host}:
    get:
      tags:
      - router-controller
      summary: "Updates the desired IP addresses of the 'host'. If both parameters\
        \ for IP addresses aren't set, an attempt is made to fetch the remote IP.\
        \ It is an alternative route for routers which requires the GET method!  The\
        \ authentication differs from the other routes. The user is the host and the\
        \ password is the 'apiToken'."
      operationId: routerUpdateHost
      parameters:
      - name: host
        in: path
        description: "The host, for which the IPs must be updated. It has to be a\
          \ full domain name."
        required: true
        schema:
          type: string
          description: "The host, for which the IPs must be updated. It has to be\
            \ a full domain name."
          example: mydyndns.domain.com
        example: mydyndns.domain.com
      - name: ipv4
        in: query
        description: An IPv4 address.
        required: false
        schema:
          type: string
          description: An IPv4 address.
          examples:
          - 127.1.2.4
      - name: ipv6
        in: query
        description: An IPv6 address.
        required: false
        schema:
          type: string
          description: An IPv6 address.
          examples:
          - 2a03:4000:41:32:0:0:0:2
      responses:
        "200":
          description: The IPs are still up to date.
        "201":
          description: 'One or both IPs are changed. Update successful processed.
            (Dependent on the configuration the response code could be ''200''! '
        "400":
          description: At least one IP address isn't valid or the remote IP couldn't
            be determined.
        "403":
          description: If the authentication is wrong.
        "404":
          description: The desired host doesn't exists.
        "500":
          description: If the addOrUpdate failed.
...